BOT/requirements.txt
aiogram==3.7.0
groq
asyncpg
python-dotenv
fastapi
uvicorn
jinja2
apscheduler